requirements.txt
datasets==2.20.0
langchain==0.2.7
langchain_anthropic==0.1.22
langchain-openai==0.1.25
langchain-mistralai==0.1.12
langchain-community==0.2.7
langchain_cohere==0.2.0
langchain_core==0.2.28
langchain_fireworks==0.1.7
numpy==1.26.1
openai==1.35.10
voyageai==0.2.3
anthropic==0.31.0
pandas==2.1.4
python-dotenv==0.21.0
PyYAML==6.0.2
ratelimit==2.2.1
scikit_learn==1.4.2
scipy==1.14.0
setuptools==69.5.1
tenacity==8.5.0
tqdm==4.66.4
umap_learn==0.5.3